pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = ""
authors = [
    {name = "goravjhabakh",email = "goravjhabakh@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "langchain (>=0.3.25,<0.4.0)",
    "openai (>=1.79.0,<2.0.0)",
    "pymongo (>=4.13.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pdfplumber (>=0.11.6,<0.12.0)"
]

[tool.poetry]
packages = [{include = "project", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
